+++ spec/stuart/streaming/ApacheSpark_StreamingContextSuite_spec.lua
@@ -2,6 +2,7 @@ local Context = require 'stuart.Context'
 local moses = require 'moses'
 moses.range = require 'stuart.util.mosesPatchedRange'
 local registerAsserts = require 'registerAsserts'
+local SparkConf = require 'stuart.SparkConf'
 local StreamingContext = require 'stuart.streaming.StreamingContext'
 local stuart = require 'stuart'
 
@@ -72,4 +73,43 @@ describe('Apache Spark 2.2.0 StreamingContextSuite', function()
     assert.equals('stopped', ssc:getState())
   end)
   
+  it('stop only streaming context', function()
+    local conf = SparkConf:new():setMaster(master):setAppName(appName)
+    
+    -- Explicitly do not stop SparkContext
+    local ssc = stuart.NewStreamingContext(conf, batchDuration)
+    local sc = ssc.sparkContext
+    --addInputStream(ssc).register()
+    ssc:start()
+    local stopSparkContext = false
+    ssc:stop(stopSparkContext)
+    assert.equal('stopped', ssc:getState())
+    assert.equal(100, #sc:makeRDD(moses.range(1, 100)):collect())
+    sc:stop()
+    
+    -- Implicitly do not stop SparkContext
+    conf:set('spark.streaming.stopSparkContextByDefault', 'false')
+    ssc = stuart.NewStreamingContext(conf, batchDuration)
+    sc = ssc.sparkContext
+    --addInputStream(ssc).register()
+    ssc:start()
+    ssc:stop()
+    assert.equal(100, #sc:makeRDD(moses.range(1, 100)):collect())
+    sc:stop()
+  end)
+
+  it('stop(stopSparkContext=true) after stop(stopSparkContext=false)', function()
+    local ssc = stuart.NewStreamingContext(master, appName, batchDuration)
+--    addInputStream(ssc).register()
+    local stopSparkContext = false
+    ssc:stop(stopSparkContext)
+    assert.equal(100, #ssc.sc:makeRDD(moses.range(1, 100)):collect())
+    stopSparkContext = true
+    ssc:stop(stopSparkContext)
+    -- Check that the SparkContext is actually stopped:
+    assert.has_error(function()
+      ssc.sc:makeRDD(moses.range(1, 100)):collect()
+    end)
+  end)
+
 end)